fix: skipping layout because adapter is not set
- fixed two of the same errors that were logged (not seen by user) where
the view did not have an adapter set (which was causing the layout to be
"skipped"). error was logged adb. fix was to make sure that an initial
adapter was set in the view create method. previously, it was only being
assigned after the view was created and the data for the adapter was
being retrieved.